+++ room_reservation/api/admin.py
@@ -1,3 +1,8 @@
 from django.contrib import admin
 
-# Register your models here.
+from .models import Participant, Reservation, Room
+
+
+@admin.register(Room, Reservation, Participant)
+class MultipleAdmin(admin.ModelAdmin):
+    pass
