--- room_reservation/reservations/admin.py
@@ -1,8 +1,0 @@
-from django.contrib import admin
-
-from .models import Participant, Reservation
-
-
-@admin.register(Reservation, Participant)
-class ReservationsAdmin(admin.ModelAdmin):
-    pass
