--- room_reservation/reservations/urls.py
@@ -1,14 +1,0 @@
-from django.urls import include, path
-from reservations import views
-from rest_framework.routers import DefaultRouter
-
-router = DefaultRouter()
-router.register(r"", views.ReservationsViewSet, basename="reservations")
-
-participants_router = DefaultRouter()
-participants_router.register(r"", views.ParticipantsViewSet, basename="participants")
-
-urlpatterns = [
-    path("", include(router.urls)),
-    path("<str:reservation_id>/participants/", include(participants_router.urls)),
-]
