+++ room_reservation/room_reservation/settings.py
@@ -34,9 +34,6 @@ INSTALLED_APPS = [
     "corsheaders",
     "rest_framework",
     "rest_framework.authtoken",
-    "rooms",
-    "reservations",
-    "users",
     "api",
 ]
