--- room_reservation/rooms/admin.py
@@ -1,8 +1,0 @@
-from django.contrib import admin
-
-from .models import Room
-
-
-@admin.register(Room)
-class RoomAdmin(admin.ModelAdmin):
-    pass
