--- room_reservation/rooms/serializers.py
@@ -1,9 +1,0 @@
-from rest_framework import serializers
-
-from .models import Room
-
-
-class RoomSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Room
-        fields = ["id", "address", "room_name", "human_capacity"]
