--- room_reservation/rooms/urls.py
@@ -1,10 +1,0 @@
-from django.urls import include, path
-from rest_framework.routers import DefaultRouter
-from rooms import views
-
-router = DefaultRouter()
-router.register(r"", views.RoomsViewSet, basename="rooms")
-
-urlpatterns = [
-    path("", include(router.urls)),
-]
