--- room_reservation/users/serializers.py
@@ -1,15 +1,0 @@
-from django.contrib.auth import get_user_model
-from rest_framework import serializers
-from rest_framework.authtoken.models import Token
-
-
-class UserSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = get_user_model()
-        fields = ("id", "username", "email", "password")
-        extra_kwargs = {"password": {"write_only": True, "required": True}}
-
-    def create(self, validated_data):
-        user = get_user_model().objects.create_user(**validated_data)
-        Token.objects.create(user=user)
-        return user
