--- room_reservation/users/urls.py
@@ -1,10 +1,0 @@
-from django.urls import include, path
-from rest_framework.routers import DefaultRouter
-from users import views
-
-router = DefaultRouter()
-router.register(r"", views.UsersViewSet, basename="users")
-
-urlpatterns = [
-    path("", include(router.urls)),
-]
