--- room_reservation/users/views.py
@@ -1,11 +1,0 @@
-from django.contrib.auth import get_user_model
-from rest_framework import viewsets
-from rest_framework.permissions import AllowAny
-
-from .serializers import UserSerializer
-
-
-class UsersViewSet(viewsets.ModelViewSet):
-    queryset = get_user_model().objects.all()
-    permission_classes = [AllowAny]
-    serializer_class = UserSerializer
